Loopdecorators now dont restart automatically
UntilFail and UntilSucceed now only does a rerun on BT.step() instead of
getting stuck in a while loop like before. Not sure if this is the best
solution generally, but works best for our use since we end up using way
too much memory when a bad tree gets generated.